Déplace les fonctions basiques.
